refactor and cleanup after code review
